Add files via upload
This is Siddhi Vehicle Rental's main page.
On the customer tab, the page displays the bookings made by existing clients. The user can create a new one at the bottom of the page by selecting one from the drop-down lists and inputting the date and time for when they wish to book.